From Michael Homer:
"updated recipe for HAL that works with newer versions
of PCIUtils. It adds "LDFLAGS=-lz" to the environment, as that's now a
requirement of PCIUtils>=2.2.4. I imagine the next release of HAL will
handle this itself, so it's only an interim measure.

It also adds file_size and MD5 sums, and allows HAL to install its
unmanaged files in /S/V/run/hald."