dnsjava is a dns library written completely in java. this recipe supplies the binary version.  if I can get the source verion to properly compile under makefile I'll get that up here.
I've provided a couple of simple bash scripts to test for a 
java installation and then run a couple of bundled commandline 
tools that are part of the jar.  there's a dig and lookup 
tool that pull information from the associated dns table.

If anyone finds an issue with this package, please contact 
me:  nathan dot middleton at gmail dot com