According to Jonas:
The [Scripts] implementation is changed so that options are passed verbatim to the responsible script, i.e. "--conflict overwrite" is passed as _one_ string. This is so that arguments can contain spaces. Adding a equal sign will work, since Parse_Options will trigger on that and split the string into option and argument.