Fixes outstanding problems with ecryptfs on rootfs
This patch relies on an updated snapshot of ConfigTools (so that the
changes made by WriteBoot64 to GRUB_CMDLINE_LINUX are not lost). It
also needs a modified version of Dracut (we'll be sending our patches
upstream soon) so that it correctly parses the cryptdevice= argument
even in the absence of systemd.