Linux: read version from include/config/kernel.release
The original recipe attempted to define the kernel version by combining
the recipe's own version with CONFIG_LOCALVERSION, but that's prone to
failures -- especially when the kernel is build more than once (i.e., a
suffix -2, -3, etc is added to the version string) and when building
from a git snapshot (such as linux-next).

We now read the version from include/config/kernel.release so that both
`depmod` and `WriteBoot64` look for modules under the correct
`lib/modules/version` directory. This should fix the problem reported at
https://gobolinux.discourse.group/t/installation-of-linux-kernel-fails/38/4